fix: generate readme's Configuration section based on the action model
