+++ packages/adk/lib/bootstrap/generators/license.ts
@@ -4,10 +4,11 @@ import { Injectable, Logger, Scope } from '@nestjs/common';
 import { BootstrapGenerator, BootstrapGeneratorResult, BootstrapError, GeneratorProps } from '../model';
 import { Model } from '@aws/codecatalyst-adk-model-parser';
 import { writeContentToFileSync } from '@aws/codecatalyst-adk-utils/lib';
+import { read_space } from '../../util/space';
 
 
 export const LICENSE_GENERATOR = 'license_generator';
-export const ACTION_CONFIG_FILE = '.actionconfig';
+
 export const LICENSE_GENERATOR_DESTINATION_FILES = ['LICENSE'];
 
 @Injectable({ scope: Scope.DEFAULT })
@@ -17,20 +18,8 @@ export class LicenseGenerator implements BootstrapGenerator {
         try {
             Logger.log('Generating MIT LICENSE');
             const year = new Date().getFullYear();
-            let copyRightHolder = '';
-
             const templateContents = fs.readFileSync(`${props.templateBasePath}/templates/action/${props.language}/LICENSE.md`, 'utf-8');
-            if (fs.existsSync(ACTION_CONFIG_FILE)) {
-                const actionConfig = fs.readFileSync(ACTION_CONFIG_FILE, 'utf-8');
-                if (actionConfig) {
-                    actionConfig.split(/\r?\n/).forEach(function (line) {
-                        const space = line.split('codecatalyst_space: ')[1];
-                        if (space) {
-                            copyRightHolder = space;
-                        }
-                    });
-                }
-            }
+            const copyRightHolder = read_space();
 
             let templateKeys: { [key: string]: string } = {
                 YEAR: `${year}`,
